+++ src/main/kotlin/com/github/sleepypanda/feesh/settings/categories/WorldRendering.kt
@@ -1,5 +1,6 @@
 package com.github.sleepypanda.feesh.settings.categories
 
+import com.github.sleepypanda.feesh.constants.ModVersionConstants
 import com.github.sleepypanda.feesh.utils.enums.ColorCodes.*
 import com.github.sleepypanda.feesh.utils.enums.FormattingCodes.*
 import com.github.sleepypanda.feesh.features.rendering.HidePlayersNearBobber
@@ -87,6 +88,7 @@ object WorldRendering : CategoryKt("World Rendering") {
     var hideTadgangNametags by boolean(false) {
         this.name = Translated("Hide Tadgang nametags")
         this.description = Translated("Hides Tadgang tadpoles nametags in Moonglade Marsh. Tadgang frogs nametags are still visible!.")
+        this.searchTerms = listOf(ModVersionConstants.VERSION_1_12_0)
     }
 
     init {
@@ -97,11 +99,11 @@ object WorldRendering : CategoryKt("World Rendering") {
 
     var muteJadeDragon by boolean(false) {
         this.name = Translated("Mute Jade Dragon")
-        this.description = Translated("Mutes Jade dragon sounds while you are in dragon's cave.")
+        this.description = Translated("Mutes Jade dragon sounds 'entity.ender_dragon.*' while you are in dragon's cave. ${YELLOW}If you use Sound Controller mod, this setting might be overridden by it!")
     }
 
     var muteReindrakeGifts by boolean(false) {
         this.name = Translated("Mute Reindrake gifts")
-        this.description = Translated("Mutes loud 'totem used' sounds while picking up gifts from own Reindrake.")
+        this.description = Translated("Mutes loud 'item.totem.use' sounds while picking up gifts from a Reindrake. ${YELLOW}If you use Sound Controller mod, this setting might be overridden by it!")
     }
 }